athome/requirements.txt
pygubu==0.14
pycryptodome